Refactor parser dir
