+++ src/Makefile.am
@@ -87,10 +87,10 @@ souffle_sources = \
         GraphUtils.h                                       \
         IOSystem.h                                         \
         LogStatement.h                                     \
-        ParserDriver.cpp                                   \
-        ParserDriver.h                                     \
-        ParserUtils.cpp                                    \
-        ParserUtils.h                                      \
+        parser/ParserDriver.cpp                            \
+        parser/ParserDriver.h                              \
+        parser/ParserUtils.cpp                             \
+        parser/ParserUtils.h                               \
         ProfileEvent.h                                     \
         RamTypes.h                                         \
         ReadStream.h                                       \
@@ -270,8 +270,8 @@ souffle_sources = \
         interpreter/InterpreterProgInterface.h             \
         interpreter/InterpreterRelation.cpp                \
         interpreter/InterpreterRelation.h                  \
-        parser.cc                                          \
-        parser.hh                                          \
+        parser/parser.cc                                   \
+        parser/parser.hh                                   \
         ram/Condition.h                                    \
         ram/Expression.h                                   \
         ram/Node.h                                         \
@@ -318,8 +318,8 @@ souffle_sources = \
         ram/transform/ReportIndex.h                        \
         ram/transform/TupleId.cpp                          \
         ram/transform/TupleId.h                            \
-        scanner.cc                                         \
-        stack.hh                                           \
+        parser/scanner.cc                                  \
+        parser/stack.hh                                    \
         synthesiser/Synthesiser.cpp                        \
         synthesiser/Synthesiser.h                          \
         synthesiser/SynthesiserRelation.cpp                \
@@ -342,7 +342,7 @@ souffle_profile_SOURCES = souffle_prof.cpp
 
 dist_bin_SCRIPTS = souffle-compile souffle-config
 
-EXTRA_DIST = parser.yy scanner.ll
+EXTRA_DIST = parser/parser.yy parser/scanner.ll
 
 soufflepublicdir = $(includedir)/souffle
 
@@ -397,17 +397,17 @@ souffleutilitydir = $(soufflepublicdir)/utility
 souffleutility_HEADERS = $(souffle_utility_sources)
 
 # files to clean
-CLEANFILES = $(BUILT_SOURCES)  parser.cc scanner.cc parser.hh stack.hh
+CLEANFILES = $(BUILT_SOURCES)  parser/parser.cc parser/scanner.cc parser/parser.hh parser/stack.hh
 
 # run Bison
-$(builddir)/parser.hh: $(srcdir)/parser.yy
-	$(BISON) -Wall -Werror -Wno-error=deprecated -Wno-error=other -v -d -o parser.cc $(srcdir)/parser.yy
+$(builddir)/parser/parser.hh: $(srcdir)/parser/parser.yy
+	$(BISON) -Wall -Werror -Wno-error=deprecated -Wno-error=other -v -d -o parser/parser.cc $(srcdir)/parser/parser.yy
 
 # and FLEX
-$(builddir)/scanner.cc: $(srcdir)/scanner.ll
-	$(FLEX) -o scanner.cc $(srcdir)/scanner.ll
+$(builddir)/parser/scanner.cc: $(srcdir)/parser/scanner.ll
+	$(FLEX) -o parser/scanner.cc $(srcdir)/parser/scanner.ll
 
 # driver depends on the generated header
-$(builddir)/parser.cc $(builddir)/stack.hh $(builddir)/scanner.cc \
-    $(builddir)/main.cpp $(builddir)/ParserDriver.cpp: $(builddir)/parser.hh
+$(builddir)/parser/parser.cc $(builddir)/parser/stack.hh $(builddir)/parser/scanner.cc \
+    $(builddir)/main.cpp $(builddir)/parser/ParserDriver.cpp: $(builddir)/parser/parser.hh
