+++ src/main.cpp
@@ -19,7 +19,6 @@
 #include "ErrorReport.h"
 #include "Explain.h"
 #include "Global.h"
-#include "ParserDriver.h"
 #include "RamTypes.h"
 #include "ast/Node.h"
 #include "ast/Program.h"
@@ -65,6 +64,7 @@
 #include "config.h"
 #include "interpreter/InterpreterEngine.h"
 #include "interpreter/InterpreterProgInterface.h"
+#include "parser/ParserDriver.h"
 #include "profile/Tui.h"
 #include "ram/Node.h"
 #include "ram/Program.h"
