+++ src/parser/ParserDriver.h
@@ -21,7 +21,7 @@
 #include "SrcLocation.h"
 #include "ast/QualifiedName.h"
 #include "ast/TranslationUnit.h"
-#include "parser.hh"
+#include "parser/parser.hh"
 #include <cstdio>
 #include <memory>
 #include <set>
