+++ src/tests/ast_parser_utils_test.cpp
@@ -16,7 +16,7 @@
 
 #include "tests/test.h"
 
-#include "ParserUtils.h"
+#include "parser/ParserUtils.h"
 #include "ast/Atom.h"
 #include "ast/Clause.h"
 #include "ast/Node.h"
